removed inline styles and added to css file
I needed better control over the layout of the module. Found much was hardcoded into the module's .js file so I removed it and added class names for logical layout such as title, image, descrption, and copyright. I also added to the module's css file. I need to clean that up some as there is extra junk in there that is not used.